% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{projr_init}
\alias{projr_init}
\title{Initialise a New Project with Standard Directories, Git, and README}
\usage{
projr_init(
  dir_raw_data = "_raw_data",
  dir_cache = "_tmp",
  dir_docs = "docs",
  dir_output = "_output",
  dir_reference = "_reference",
  init_git = TRUE,
  git_add_and_commit_all = TRUE,
  init_readme = TRUE,
  init_github = TRUE,
  github_private = TRUE
)
}
\arguments{
\item{dir_raw_data}{Character. Directory for raw data. Defaults to
\code{"_raw_data"}. Set to \code{NULL} to skip creating this directory.}

\item{dir_cache}{Character. Directory for temporary or cache files. Defaults to
\code{"_tmp"}. Set to \code{NULL} to skip.}

\item{dir_docs}{Character. Directory for rendered documents. Defaults to
\code{"docs"}. Set to \code{NULL} to skip.}

\item{dir_output}{Character. Directory for final outputs. Defaults to
\code{"_output"}. Set to \code{NULL} to skip.}

\item{dir_reference}{Character. Directory for non-rendered docs. Defaults to
\code{"_reference"}. Set to \code{NULL} to skip.}

\item{init_git}{Logical. If \code{TRUE} (the default), initialises a Git repository
in the current project.}

\item{git_add_and_commit_all}{Logical. If \code{TRUE} (the default), attempts to add
and commit all untracked and modified files after Git initialisation.}

\item{init_readme}{Logical. If \code{TRUE} (the default), creates a README file with
recommended project instructions.}

\item{init_github}{Logical. If \code{TRUE}, connects the project to GitHub via
\code{usethis::use_github()}. Defaults to \code{TRUE}.}

\item{github_private}{Logical. If \code{TRUE} (the default) and if GitHub is
initialised (i.e. when \code{init_github = TRUE}), the GitHub repository is created
as private. A message is issued to remind the user that collaborators must be
added manually via GitHub settings.}
}
\value{
Invisibly returns \code{TRUE} when complete.
}
\description{
\code{projr_init} creates a standard project structure by making several
directories for raw data, temporary/cache files, final outputs, rendered docs,
and raw (non-rendered) docs. It also can initialise a Git repository (adding
the cache directory to \code{.gitignore} and optionally adding & committing all
changes), write a README file with recommended project instructions, and, if
desired, connect the project to GitHub.
}
\details{
The default behaviour is to create all standard directories, initialise Git
(adding the cache directory to \code{.gitignore} and committing changes), and write a
README. If you do not wish to create one or more of the directories, pass
\code{NULL} for that parameter.
}
\examples{
\dontrun{
projr_init(
  dir_raw_data = "_raw_data",
  dir_cache = "_tmp",
  dir_docs = "docs",
  dir_output = "_output",
  dir_reference = "_reference",
  init_git = TRUE,
  git_add_and_commit_all = TRUE,
  init_readme = TRUE,
  init_github = TRUE,
  github_private = TRUE
)
}
}
