% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{projr_run}
\alias{projr_run}
\title{Run project scripts, optionally clearing output and copying docs}
\usage{
projr_run(
  scripts = NULL,
  skip_quarto_project = FALSE,
  clear_output_and_docs = FALSE,
  copy_docs = TRUE,
  dir_scripts = NULL,
  dir_output = NULL,
  dir_docs = NULL,
  dir_exec = NULL
)
}
\arguments{
\item{scripts}{Character vector of script filenames, e.g.
\code{c("my_script.R", "my_doc.Rmd")}. If \code{NULL}, all \code{.R}, \code{.Rmd}/\code{.rmd}, and
\code{.qmd} files in \code{dir_scripts} are used.}

\item{skip_quarto_project}{Logical. If \code{TRUE}, does not render the Quarto
project, even if \verb{_quarto.yml} is found.}

\item{clear_output_and_docs}{Logical. If \code{TRUE}, clears out \code{dir_output} and
\code{dir_docs} before running.}

\item{copy_docs}{Logical. If \code{TRUE}, copies generated output files (\code{.html},
\code{.pdf}, \code{.docx}) into \code{dir_docs}. Does not apply to Quarto projects.}

\item{dir_scripts}{Directory where scripts reside. Defaults to the working
directory if \code{NULL}.}

\item{dir_output}{Directory to clear if \code{clear_output_and_docs = TRUE}.
Defaults to \code{"_output"} if \code{NULL}.}

\item{dir_docs}{Directory where final docs are copied. Defaults to \code{"docs"}
if \code{NULL}.}
}
\value{
Invisibly returns \code{TRUE} when complete.
}
\description{
This function:
\enumerate{
\item Collects R, Rmd, or qmd scripts from \code{dir_scripts}, or uses a supplied
character vector.
\item Optionally clears \code{dir_output} and \code{dir_docs}.
\item Sources or renders each script, skipping \code{.qmd} scripts if a Quarto project
is detected.
\item By default, copies generated \code{.html}, \code{.pdf}, or \code{.docx} files to
\code{dir_docs}.
\item By default, renders the entire Quarto project if \verb{_quarto.yml} is found.
}
}
\details{
To avoid accidental deletion of outputs,
\code{clear_output_and_docs} is by default \code{FALSE}.
However, for a clean run, it is recommended to set \code{clear_output_and_docs}
to \code{TRUE}.
}
\examples{
\dontrun{
# Run all scripts in "scripts", remove old docs/outputs, and copy
# newly generated docs to "docs"
projr_run(
  skip_quarto_project = TRUE,
  clear_output_and_docs = TRUE,
  copy_docs = TRUE
)
}
}
